requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
prophet==1.1.5
sentence-transformers==2.2.2
nltk>=3.8.1
python-multipart==0.0.6
aiofiles==23.2.1
PyPDF2==3.0.1
python-docx==1.1.0
pydantic==2.5.0
pydantic-settings==2.1.0

